Update the Introduction to take almost all of Andy Oram's advice.
Add a paragraph (the second-to-last one before the "History" section)
about the benefits of open source, since the preceding material
discussed mostly pitfalls and extra up-front work required.

Make some corresponding tweaks to the Preface, and add a todo comment
in Chapter 8 about expanding on the new benefits paragraph above.

This commit is dedicated to Roland Tanglao, a generous backer of the
Kickstarter campaign that funded this rewrite.  Roland, thank you!


git-svn-id: http://svn.red-bean.com/repos/producingoss/trunk@2977 19bf594e-a5d2-0310-ae51-ac0f9f5f197f