Handle Andy's edits for Chapter 2.
- Eliminate the whole first half of the introduction; it doesn't
  really hold up anymore, although it was true ten years ago.

- New sidebar: "What We Mean by Users and Developers."

- Add material about converting people from user to developer.

- In "But First, Look Around", don't list SourceForge anymore.

- Mention readthedocs.org.

- Various minor updates and wording/phrasing tweaks throughout.

Archive Andy's notes in todo.org_archive, with responses from me.

This commit is for Keith Casey.  Keith, thanks so much for backing the
rewrite, and for your recent question on Kickstarter asking for a
status update.  I responded there, but I hope this commit is also a
good response.


git-svn-id: http://svn.red-bean.com/repos/producingoss/trunk@2980 19bf594e-a5d2-0310-ae51-ac0f9f5f197f