Fix the "upload" rule to just upload a PDF.
* Makefile
  (LANGUAGES): Update comment to explain.
  (upload): Just upload the PDF; don't try to do anything else.

Now it's much easier to keep the online copy of the PDF up-to-date
even if the web server's automated build infrastructure is on the
fritz, which, because this is the DocBook Lite toolchain, it
practically always is.

This commit is for Matt Soar.  Matt, thanks for pledging to the
Kickstarter campaign that made the 2nd edition possible.


git-svn-id: http://svn.red-bean.com/repos/producingoss/trunk@3004 19bf594e-a5d2-0310-ae51-ac0f9f5f197f