Make all live URL links look normal in the PDF.
This sweeping change ensures that live-link URLs are displayed in a
reasonable way in the PDF, instead of in a distracting, bad way.

The problem (arguably a bug in the DocBook->PDF conversion process,
which could be smarter about this) is that if you have something like
the following in the XML source:

  <ulink url="https://example.com/">example.com</ulink>

then it will be rendered like this in the PDF:

  example.com [https://example.com/]

(with the square braces really there, and the text in them linked).

This, of course, is rather repetitive and annoying for the reader.
BUT, there's an out: if the value of the 'url' attribute and the link
text match precisely, then the Right Thing happens:

  <ulink url="https://example.com/">https://example.com/</ulink>

becomes:

  https://example.com/

(as a live link).  The trick is that the attribute value and the text
must match exactly, character for character.  The converter will do no
intelligent guessing.  Its rule is: if they're exactly the same, then
fold them together and make the result a live link; otherwise, put the
'url' attribute value in square braces as a live link, after the text.

* en/ch00.xml, en/ch01.xml, en/ch02.xml, en/ch03.xml, en/ch04.xml,
  en/ch05.xml, en/ch06.xml, en/ch07.xml, en/ch08.xml, en/ch09.xml:
  Do the needful.

Greg Grossmeier, this commit is for you!  Thanks for backing the
Kickstarter campaign that made the 2nd edition of the book possible.


git-svn-id: http://svn.red-bean.com/repos/producingoss/trunk@3005 19bf594e-a5d2-0310-ae51-ac0f9f5f197f