Use my style of em-dash everywhere.
I prefer a space on each side of the dash, on the theory that an
em-dash should be less connective than mere space would be, and when
the dash directly touches the non-whitespace on either side it looks
deceptively more connective.  ORM house style is to use no spaces, and
for an ORM edition I can easily switch it back.  But for now, this is
my house, and I'll eat peanut butter in the living room if I want to.

* en/ch00.xml, en/ch01.xml, en/ch02.xml, en/ch03.xml, en/ch04.xml,
  en/ch05.xml, en/ch06.xml, en/ch07.xml, en/ch08.xml, en/ch09.xml: Do it.

This commit is dedicated to Camille Acey, whom I hope to get to work
with again some day.  Camille, thank you for backing the Kickstarter
campaign for the 2nd edition.  You all should come visit Chicago!
(It's lovely this time of year.  Er, don't look at the datestamp on
this commit.)


git-svn-id: http://svn.red-bean.com/repos/producingoss/trunk@3006 19bf594e-a5d2-0310-ae51-ac0f9f5f197f