Finally make conditional output work.
As noted in r2749, the <phrase output="printed">...</phrase> trick
never actually worked.  Instead, the way to do this in DocBook is
called "profiling" (http://www.sagehill.net/docbookxsl/Profiling.html)
and it only requires an expert consultant to volunteer their time to
help set everything straight.  Fortunately, that's exactly what
happened, in the person of Camille Bégnis of http://neodoc.biz/, in
#docbook in freenode IRC today.

Now we use the 'condition' attribute, with the value set to "treeware"
when we want the phrase to show up, and to "none" otherwise.  The
result is that in chunk HTML and regular (single-file) HTML output,
<phrase condition="treeware" >foo</phrase> results in "foo" being
excluded from the result, as desired, and in PDF output, the same code
results in "foo" being included.

You might think this is because everything is now working as designed,
but no, you'd be wrong.  This is the DocBook toolchain we're dealing
with; had you forgotten?  It turns out that the change made in this
commit to tools/fo-stylesheet.xsl doesn't actually have any effect,
but the changes to chunk-stylesheet.xsl html-stylesheet.xsl do, and
fortunately the direction of breakage is such that the right result
happens anyway: the text is *excluded* in both HTML outputs (because
the exclusion code is working correctly), and *included* in the PDF
(the inclusion condition is broken, but the default behavior when a
condition attribute is undefined is to assume the condition is true).

So there's still a bug here, but it's latent for now.  I tried to put
"<!-- comments about the brokenness -->" in tools/fo-stylesheet.xsl,
but having those comments then seemed break the PDF build too (the
build process hung forever once it got to the PDF stage).  I don't
know what's up with that, but also don't feel like debugging it, so
what should be in comments is instead here in this log message.

* tools/chunk-stylesheet.xsl,
  tools/html-stylesheet.xsl,
  tools/fo-stylesheet.xsl: Define profile.condition to distinguish
    between HTML output and PDF output.

* en/ch02.xml, en/ch03.xml, en/ch04.xml, en/ch05.xml, 
  en/ch06.xml, en/ch07.xml, en/ch08.xml, en/ch09.xml: Update all of the
    conditional phrases to use the new format.

* en/ch00.xml: Same, and thank Camille Bégnis in the acknowledgements.

* en/ch01.xml: No change, because no conditional phrases in that
    chapter.  Noting it here so it doesn't look like an oversight.

* todo.org, todo.org_archive: Update accordingly.

This commit is dedicated to Simon Fondrie-Teitler, a generous backer
of the Kickstarter campaign that funded the 2nd edition of this book.


git-svn-id: http://svn.red-bean.com/repos/producingoss/trunk@3025 19bf594e-a5d2-0310-ae51-ac0f9f5f197f