Give the book a version number that increments with every change.
Use the SVN rev of the most recently changed XML file as a minor
version number that comes after the major number (currently "2"),
separated by a dot as usual.

* aggrevision: New script to figure out the right SVN rev.

* Makefile
  (version-stamp): New rule.
  (LANGUAGES, all): Depend on the new rule.

* en/: Update the svn:ignore property to ignore "book.xml".

* en/book.xml: Move to...

* en/book.xml.in: ...here.  Update to invite the new version number.
  Link to the book's home site.  Don't link to the CC-BY-SA license;
  the link was distracting and not strictly necessary.

* en/copy.xml: Link to CC-BY-SA license home page from here instead.

This commit is for Tor Helmer, one of the backers of the Kickstarter
campaign that supported the 2nd edition rewrite.  Thank you, Tor.


git-svn-id: http://svn.red-bean.com/repos/producingoss/trunk@3076 19bf594e-a5d2-0310-ae51-ac0f9f5f197f